+++ src/interfaces/IState.sol
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: EVVM-NONCOMMERCIAL-1.0
+// Full license terms available at: https://www.evvm.info/docs/EVVMNoncommercialLicense
+pragma solidity ^0.8.0;
+
+interface IUserValidator {
+    function canExecute(address user) external view returns (bool);
+}